Libft

+++ libft/ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: chanhapa <chanhapa@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 22:51:25 by chanhapa          #+#    #+#             */
-/*   Updated: 2022/03/24 14:35:24 by chanhapa         ###   ########.fr       */
+/*   Updated: 2022/03/25 21:08:17 by chanhapa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*us1;
 	unsigned char	*us2;
